#908 refactor: extract user preferences into feature slice
