+++ src/components/watchlist/__tests__/WatchlistResults.test.tsx
@@ -32,7 +32,7 @@ vi.mock("lottie-react", () => ({
     default: () => <div data-testid="lottie-animation" />,
 }));
 
-vi.mock("@/hooks/preferences/useUserPreferences.tsx", () => ({
+vi.mock("@/features/preferences/hooks/useUserPreferences.tsx", () => ({
     useUserPreferences: () => ({ preferences: { currency: "EUR" }, updatePreferences: vi.fn() }),
 }));
