+++ src/features/preferences/components/CurrencySelector.tsx
@@ -4,7 +4,7 @@ import {
     parseCurrency,
     type Currency,
 } from "@/data/internal/common/Currency.ts";
-import { useUserPreferences } from "@/hooks/preferences/useUserPreferences.tsx";
+import { useUserPreferences } from "@/features/preferences/hooks/useUserPreferences.tsx";
 import { useUpdateUserAccount } from "@/features/account-management/hooks/usePatchUserAccount.ts";
 import { useUserAccount } from "@/features/account-management/hooks/useUserAccount.ts";
 import { useResolvedAuth } from "@/features/authentication/hooks/useResolvedAuth.ts";
