+++ src/hooks/search/__tests__/useSearch.test.ts
@@ -38,7 +38,7 @@ vi.mock("@/env.ts", () => ({
     },
 }));
 
-vi.mock("@/hooks/preferences/useUserPreferences.tsx", () => ({
+vi.mock("@/features/preferences/hooks/useUserPreferences.tsx", () => ({
     useUserPreferences: () => ({ preferences: { currency: "EUR" }, updatePreferences: vi.fn() }),
 }));
