+++ src/routes/__root.tsx
@@ -24,10 +24,10 @@ import "@/lib/polyfills/url";
 import "@/amplify-config.ts";
 import "@/api-config.ts";
 import { googleAnalytics } from "@/lib/tracking/googleAnalytics.ts";
-import { UserPreferencesProvider } from "@/hooks/preferences/useUserPreferences.tsx";
-import { getServerPreferences } from "@/lib/server/preferences.ts";
+import { UserPreferencesProvider } from "@/features/preferences/hooks/useUserPreferences.tsx";
+import { getServerPreferences } from "@/features/preferences/server/preferences.ts";
 import { getServerTimezone } from "@/lib/server/timezone.ts";
-import type { UserPreferences } from "@/data/internal/preferences/UserPreferences.ts";
+import type { UserPreferences } from "@/features/preferences/types/UserPreferences.ts";
 import { useTranslation } from "react-i18next";
 import { getPreferredLocale } from "@/lib/server/i18n.ts";
 import i18n from "@/i18n/i18n.ts";
